+++ docs/search/classes_d.js
@@ -6,5 +6,6 @@ var searchData=
   ['resourceidgenerator',['ResourceIdGenerator',['../interfaceBristolSU_1_1Support_1_1Authentication_1_1Contracts_1_1ResourceIdGenerator.html',1,'BristolSU::Support::Authentication::Contracts']]],
   ['revisionserviceprovider',['RevisionServiceProvider',['../classBristolSU_1_1Support_1_1Revision_1_1RevisionServiceProvider.html',1,'BristolSU::Support::Revision']]],
   ['rolefilter',['RoleFilter',['../classBristolSU_1_1Support_1_1Filters_1_1Contracts_1_1Filters_1_1RoleFilter.html',1,'BristolSU::Support::Filters::Contracts::Filters']]],
-  ['rolehasposition',['RoleHasPosition',['../classBristolSU_1_1Support_1_1Filters_1_1Filters_1_1Role_1_1RoleHasPosition.html',1,'BristolSU::Support::Filters::Filters::Role']]]
+  ['rolehasposition',['RoleHasPosition',['../classBristolSU_1_1Support_1_1Filters_1_1Filters_1_1Role_1_1RoleHasPosition.html',1,'BristolSU::Support::Filters::Filters::Role']]],
+  ['roletagged',['RoleTagged',['../classBristolSU_1_1Support_1_1Filters_1_1Filters_1_1Role_1_1RoleTagged.html',1,'BristolSU::Support::Filters::Filters::Role']]]
 ];